Add an escapeBackslash knob to Copy.expand
The underlying SimpleTemplateEngine is eager to interpret escape
sequences found in the processes files and replace them with actual
symbols, e.g. converting `\n` to newline. This is problematic for some
use cases.

This CL exposes SimpleTemplateEngine.setEscapeBackslash that allows to
turn off this behavior. Most of the changes there are just plumbing to
deliver ObjectFactory to DefaultCopySpec so it can create ExpandDetails
with a property.

Fixes #10492